+++ Assets/Scripts/Protocol/Message/Battle/FrameWeaponDropReqMsg.cs
@@ -7,12 +7,12 @@ using ZeroFrame.Buffer;namespace Game.Protocol.Battle
     public class FrameWeaponDropReqMsg
 :IZeroMessage<FrameWeaponDropReqMsg>{
         public ushort entityId; //被丢弃武器entityId
-        public byte masterId;
+        public int masterId;
 
         public void FromBytes(byte[] src, ref int offset)
         {
             entityId = BufferReader.ReadUInt16(src, ref offset);
-            masterId = BufferReader.ReadByte(src, ref offset);
+            masterId = BufferReader.ReadInt32(src, ref offset);
             offset += src.Length;
         }
 
@@ -21,7 +21,7 @@ using ZeroFrame.Buffer;namespace Game.Protocol.Battle
             int offset = 0;
             byte[] result = new byte[1000];
             BufferWriter.WriteUInt16(result, entityId, ref offset);
-            BufferWriter.WriteByte(result, masterId, ref offset);
+            BufferWriter.WriteInt32(result, masterId, ref offset);
             return result;
         }
