+++ Assets/Scripts/Protocol/Message/Battle/FrameWeaponReloadReqMsg.cs
@@ -6,11 +6,11 @@ using ZeroFrame.Buffer;namespace Game.Protocol.Battle
     [ZeroMessage]
     public class FrameWeaponReloadReqMsg
 :IZeroMessage<FrameWeaponReloadReqMsg>{
-        public byte masterId;
+        public int masterId;
 
         public void FromBytes(byte[] src, ref int offset)
         {
-            masterId = BufferReader.ReadByte(src, ref offset);
+            masterId = BufferReader.ReadInt32(src, ref offset);
             offset += src.Length;
         }
 
@@ -18,7 +18,7 @@ using ZeroFrame.Buffer;namespace Game.Protocol.Battle
         {
             int offset = 0;
             byte[] result = new byte[1000];
-            BufferWriter.WriteByte(result, masterId, ref offset);
+            BufferWriter.WriteInt32(result, masterId, ref offset);
             return result;
         }
